video_VQA_Benchmarking/requirement.txt
accelerate==1.5.2
anyio @ file:///croot/anyio_1729121277521/work
argon2-cffi @ file:///opt/conda/conda-bld/argon2-cffi_1645000214183/work
argon2-cffi-bindings @ file:///croot/argon2-cffi-bindings_1736182440035/work
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
async-lru @ file:///croot/async-lru_1699554519285/work
attrs @ file:///croot/attrs_1734533101012/work
av==14.2.0
babel @ file:///croot/babel_1737454360933/work
beautifulsoup4 @ file:///croot/beautifulsoup4-split_1718029820055/work
bitsandbytes==0.45.3
bleach @ file:///croot/bleach_1732290411627/work
Brotli @ file:///croot/brotli-split_1736182456865/work
certifi @ file:///croot/certifi_1738623731865/work/certifi
cffi @ file:///croot/cffi_1736182485317/work
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
comm @ file:///croot/comm_1709322850197/work
contourpy @ file:///croot/contourpy_1732540045555/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
debugpy @ file:///croot/debugpy_1736267418885/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
exceptiongroup @ file:///croot/exceptiongroup_1706031385326/work
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
fastjsonschema @ file:///croot/python-fastjsonschema_1731939362158/work
filelock @ file:///croot/filelock_1700591183607/work
fonttools @ file:///croot/fonttools_1737039080035/work
fsspec @ file:///croot/fsspec_1736274672181/work
gmpy2 @ file:///croot/gmpy2_1738085463648/work
h11 @ file:///croot/h11_1706652277403/work
httpcore @ file:///croot/httpcore_1706728464539/work
httpx @ file:///croot/httpx_1723474802858/work
huggingface-hub==0.29.2
idna @ file:///croot/idna_1714398848350/work
ipykernel @ file:///croot/ipykernel_1737660677549/work
ipython @ file:///croot/ipython_1734548052611/work
ipywidgets @ file:///croot/ipywidgets_1733504575802/work
jedi @ file:///croot/jedi_1733987392413/work
Jinja2 @ file:///croot/jinja2_1737760107953/work
json5 @ file:///croot/json5_1730786798687/work
jsonschema @ file:///croot/jsonschema_1728486696720/work
jsonschema-specifications @ file:///croot/jsonschema-specifications_1699032386549/work
jupyter @ file:///croot/jupyter_1737645803214/work
jupyter-console @ file:///croot/jupyter_console_1679999630278/work
jupyter-events @ file:///croot/jupyter_events_1741184577592/work
jupyter-lsp @ file:///croot/jupyter-lsp-meta_1699978238815/work
jupyter_client @ file:///croot/jupyter_client_1737570961872/work
jupyter_core @ file:///croot/jupyter_core_1718818295206/work
jupyter_server @ file:///croot/jupyter_server_1741200034823/work
jupyter_server_terminals @ file:///croot/jupyter_server_terminals_1686870725608/work
jupyterlab @ file:///croot/jupyterlab_1737555423487/work
jupyterlab_pygments @ file:///croot/jupyterlab_pygments_1741124142640/work
jupyterlab_server @ file:///croot/jupyterlab_server_1725865349919/work
jupyterlab_widgets @ file:///croot/jupyterlab_widgets_1733440870459/work
kiwisolver @ file:///croot/kiwisolver_1737039087198/work
MarkupSafe @ file:///croot/markupsafe_1738584038848/work
matplotlib==3.10.0
matplotlib-inline @ file:///opt/conda/conda-bld/matplotlib-inline_1662014470464/work
mistune @ file:///croot/mistune_1741124011532/work
mkl-service==2.4.0
mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
mpmath @ file:///croot/mpmath_1690848262763/work
nbclient @ file:///croot/nbclient_1741123995822/work
nbconvert @ file:///croot/nbconvert-meta_1741184653315/work
nbformat @ file:///croot/nbformat_1728049424075/work
nest-asyncio @ file:///croot/nest-asyncio_1708532673751/work
networkx @ file:///croot/networkx_1737039604450/work
notebook @ file:///croot/notebook_1738159946465/work
notebook_shim @ file:///croot/notebook-shim_1699455894279/work
numpy @ file:///croot/numpy_and_numpy_base_1708638617955/work/dist/numpy-1.26.4-cp310-cp310-linux_x86_64.whl#sha256=d8cd837ed43e87f77e6efaa08e8de927ca030a1c9c5d04624432d6fb9a74a5ee
overrides @ file:///croot/overrides_1699371140756/work
packaging @ file:///croot/packaging_1734472117206/work
pandocfilters @ file:///opt/conda/conda-bld/pandocfilters_1643405455980/work
parso @ file:///croot/parso_1733963305961/work
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pillow @ file:///croot/pillow_1738010226202/work
platformdirs @ file:///croot/platformdirs_1692205439124/work
prometheus_client @ file:///croot/prometheus_client_1731953121795/work
prompt-toolkit @ file:///croot/prompt-toolkit_1704404351921/work
protobuf==6.30.1
psutil @ file:///croot/psutil_1736367091698/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments @ file:///croot/pygments_1684279966437/work
pyparsing @ file:///croot/pyparsing_1731445506121/work
PyQt6==6.7.1
PyQt6_sip @ file:///croot/pyqt-split_1740498191142/work/pyqt_sip
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dateutil @ file:///croot/python-dateutil_1716495738603/work
python-json-logger @ file:///croot/python-json-logger_1734370021104/work
PyYAML @ file:///croot/pyyaml_1728657952215/work
pyzmq @ file:///croot/pyzmq_1734687138743/work
qtconsole @ file:///croot/qtconsole_1737590761179/work
QtPy @ file:///croot/qtpy_1700144840038/work
referencing @ file:///croot/referencing_1699012038513/work
regex==2024.11.6
requests @ file:///croot/requests_1730999120400/work
rfc3339-validator @ file:///croot/rfc3339-validator_1683077044675/work
rfc3986-validator @ file:///croot/rfc3986-validator_1683058983515/work
rpds-py @ file:///croot/rpds-py_1736541261634/work
safetensors==0.5.3
scipy==1.15.2
Send2Trash @ file:///croot/send2trash_1736540790175/work
sentencepiece==0.2.0
sip @ file:///croot/sip_1738856193618/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio @ file:///croot/sniffio_1705431295498/work
soupsieve @ file:///croot/soupsieve_1696347547217/work
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
sympy @ file:///croot/sympy_1738108488918/work
terminado @ file:///croot/terminado_1671751832461/work
tinycss2 @ file:///croot/tinycss2_1738337643607/work
tokenizers==0.21.1
tomli @ file:///opt/conda/conda-bld/tomli_1657175507142/work
torch @ file:///opt/conda/envs/bld/conda-bld/pytorch-select_1731532454354/work
torchvision @ file:///opt/conda/envs/bld/conda-bld/torchvision_1722529934174/work
tornado @ file:///croot/tornado_1733960490606/work
tqdm==4.67.1
traitlets @ file:///croot/traitlets_1718227057033/work
transformers @ git+https://github.com/huggingface/transformers.git@27361bd218e5d1ecffc68b9317c49d3a764dfbba
triton @ file:///opt/conda/envs/bld/conda-bld/torchtriton_1721056825793/work/python
typing_extensions @ file:///croot/typing_extensions_1734714854207/work
unicodedata2 @ file:///croot/unicodedata2_1736541023050/work
urllib3 @ file:///croot/urllib3_1737133630106/work
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webencodings==0.5.1
websocket-client @ file:///croot/websocket-client_1715878298792/work
widgetsnbextension @ file:///croot/widgetsnbextension_1733439572884/work
